irc: prompt for nickserv password before connect
Instead of store the IRC nickerv password as plaintext on yaml config
file let the admin write it in runtime just after the config file is
read. The input is made through an password specific input field.

Note that the password ('irc:password = ""' on config file) must be
empty, once the field is filled that data will be used as password. Also
note that this patch doesn't handle any secure transfer or storage of
the password.