config: add support to store secrets as env vars
The ability to handle secrets as env vars is useful when running
automated and continuous delivery and integration, because of that this
supporte was added using 'viper' library.

Now the user can specify the secrets (IRC nickserv password, Slack token
and Telegram tokan) in three different forms:

    1. though configuration file
    2. environment variables
    3. direct input on bot startup

If the user let some of the secret options on configuration file empty
the second option will be used and if no env vars are found the third is
prompted.